Added code file
**Business Problem**

My teammate, Zhangbo and I, both love blueberries. We have it in our pastries, our milkshakes and in our jams. This is why we were wondering what it takes to actually grow blueberries and what factors assess its yield. This also led to suggesting actionable decision variables for prospective blueberry farmers that would be able to decide where to set up their farms to get the best bang for their buck. Afterall, there can never be too many blueberries in this world.